update to match crescendo 2.2.4 where double is replaced with long to avoid rounding errors. Fixed issue with getMaxStepSize where initial steps would return large steps due to type conversion errors, now it turns double min in that case.

+++ core/fmi-interpreter/src/main/java/org/crescendo/fmi/FmiSimulationManager.java
@@ -53,7 +53,7 @@ public class FmiSimulationManager extends SimulationManager
 	 * @throws RemoteSimulationException
 	 *             thrown if an internal error occur
 	 */
-	public synchronized List<NamedValue> step(Double outputTime,
+	public synchronized List<NamedValue> step(long outputTime,
 			List<NamedValue> inputs) throws RemoteSimulationException
 	{
 		checkMainContext();
